Refactored Seasons section
